+++ src/features/auth/LoginForm.tsx
@@ -1,32 +1,37 @@
-import { ReactElement, cloneElement } from "react";
-import { useForm } from "react-hook-form";
+import TextInput from "@/components/text-input/TextInput";
+import { Control, Controller, FormState } from "react-hook-form";
 import { LoginForm as FormType } from "./AuthFormType";
-import s from "./auth.module.css";
 
-const initialValues = {
-  email: "",
-  password: "",
-};
-
-const LoginForm = ({ children }: { children: ReactElement }) => {
-  const { handleSubmit, control, formState } = useForm<FormType>({
-    defaultValues: initialValues,
-    mode: "onBlur",
-  });
-
-  const onSubmit = (data: FormType) => {
-    alert(JSON.stringify(data));
-  };
-
-  const inputs = cloneElement(children, { control: control, errors: formState.errors });
-
+const LoginForm = ({ control, errors }: { control: Control; errors: FormState<FormType>["errors"] }) => {
   return (
-    <form onSubmit={handleSubmit(onSubmit)} className={s.authForm}>
-      {inputs}
-      <button type="submit" className={s.submitBtn}>
-        Sign In
-      </button>
-    </form>
+    <>
+      <Controller
+        name="email"
+        control={control}
+        rules={{
+          required: true,
+        }}
+        render={({ field: props }) => (
+          <TextInput type="email" label="Email" autoComplete="email" error={errors?.email?.message} {...props} />
+        )}
+      />
+
+      <Controller
+        name="password"
+        control={control}
+        rules={{
+          required: true,
+          pattern: {
+            value:
+              /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[?!@#$%^&*()-_+=.,~[\]{};:'"|/><()])[A-Za-z\d?!@#$%^&*()-_+=.,~[\]{};:'"|/><()]{4,}$/,
+            message: "Ne valja",
+          },
+        }}
+        render={({ field: props }) => (
+          <TextInput type="password" label="Password" error={errors?.password?.message} {...props} />
+        )}
+      />
+    </>
   );
 };
